unit_test/generator/requirements.txt
#requirmenst.txt

kafka-python == 2.0.1
pandas == 1.0.5